requirements.txt
Flask
numpy
pandas
tensorflow-cpu
pillow
keras
Werkzeug